refactor: move message handler under log directory
