+++ vicinae/src/log/message-handler.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include <iostream>
+#include <qdatetime.h>
+#include <qlogging.h>
+#include <filesystem>
+
+void coloredMessageHandler(QtMsgType type, const QMessageLogContext &context, const QString &msg);
